app/requirements.txt
Flask
plotly
numpy
pandas
python-dotenv